Move all screens into a screen/ folder
