+++ homebase-chat/src/commonMain/kotlin/id/homebase/chat/screen/ConversationMessagesPane.kt
@@ -1,4 +1,4 @@
-package id.homebase.chat.widget
+package id.homebase.chat.screen
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.SharedTransitionLayout
